So maby changes :))

+++ Backend/src/main/java/com/studysphere/backend/service/StudentService.java
@@ -1,6 +1,6 @@
 package com.studysphere.backend.service;
 
-import com.studysphere.backend.model.Attendance;
+import com.studysphere.backend.exceptions.EmailAlreadyExistsException;
 import com.studysphere.backend.model.School;
 import com.studysphere.backend.model.people.Person;
 import com.studysphere.backend.model.people.Student;
@@ -8,14 +8,13 @@ import com.studysphere.backend.repository.AttendanceRepository;
 import com.studysphere.backend.repository.PersonRepository;
 import com.studysphere.backend.repository.SchoolRepository;
 import com.studysphere.backend.repository.StudentRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -36,9 +35,13 @@ public class StudentService {
     @Transactional
     public Student add(Student student) {
         Person person = student.getPerson();
+        if (personRepository.findByEmail(person.getEmail()).isPresent()) {
+            throw new EmailAlreadyExistsException("Email already exists: " + person.getEmail());
+        }
         personRepository.save(person);
-        School school = schoolRepository.findById(student.getSchool().getId()).orElse(null);
-        assert school != null;
+        School school = schoolRepository.findById(student.getSchool().getId())
+                .orElseThrow(() -> new EntityNotFoundException("School not found with ID: " + student.getSchool().getId()));
+        
         school.getStudentList().add(student);
         return studentRepository.save(student);
     }
@@ -48,17 +51,19 @@ public class StudentService {
     }
 
     @Transactional
-    @Modifying
     public boolean removeById(Long id) {
         System.out.println(id);
         Student student = studentRepository.findById(id).orElse(null);
-        assert student != null;
+        if (student == null) {
+            throw new EntityNotFoundException("Student not found with ID: " + id);
+        }
         School school = student.getSchool();
         school.setStudentList(school.getStudentList().stream().filter(s -> !Objects.equals(s.getId(), student.getId())).toList());
         personRepository.deleteById(student.getPerson().getId());
         studentRepository.deleteById(id);
         return false;
     }
+    //Todo test removing @Modifying
 
     public void removeAll() {
         studentRepository.deleteAll();
